+++ samples/mario-pizzeria/ui/controllers/__init__.py
@@ -4,10 +4,12 @@ from ui.controllers.delivery_controller import UIDeliveryController
 from ui.controllers.home_controller import HomeController
 from ui.controllers.kitchen_controller import UIKitchenController
 from ui.controllers.management_controller import UIManagementController
+from ui.controllers.notifications_controller import UINotificationsController
 
 __all__ = [
     "HomeController",
     "UIKitchenController",
     "UIDeliveryController",
     "UIManagementController",
+    "UINotificationsController",
 ]